requirements.txt
django
psycopg2-binary
Pillow
djangorestframework
python-dotenv